+++ src/components/HybridAppMiddleware/index.ios.tsx
@@ -3,6 +3,7 @@ import {useContext, useEffect, useState} from 'react';
 import {NativeEventEmitter, NativeModules} from 'react-native';
 import type {NativeModule} from 'react-native';
 import {useOnyx} from 'react-native-onyx';
+import type {OnyxEntry} from 'react-native-onyx';
 import {InitialURLContext} from '@components/InitialURLContextProvider';
 import useExitTo from '@hooks/useExitTo';
 import useSplashScreen from '@hooks/useSplashScreen';
@@ -14,10 +15,21 @@ import * as Welcome from '@userActions/Welcome';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type {HybridAppRoute, Route} from '@src/ROUTES';
+import type {TryNewDot} from '@src/types/onyx';
 
 type HybridAppMiddlewareProps = {
     authenticated: boolean;
     children: React.ReactNode;
+};
+
+const onboardingStatusSelector = (tryNewDot: OnyxEntry<TryNewDot>) => {
+    let completedHybridAppOnboarding = tryNewDot?.classicRedirect?.completedHybridAppOnboarding;
+
+    if (typeof completedHybridAppOnboarding === 'string') {
+        completedHybridAppOnboarding = completedHybridAppOnboarding === 'true';
+    }
+
+    return completedHybridAppOnboarding;
 };
 
 /*
@@ -36,6 +48,20 @@ function HybridAppMiddleware({children, authenticated}: HybridAppMiddlewareProps
 
     const [isAccountLoading] = useOnyx(ONYXKEYS.ACCOUNT, {selector: (account) => account?.isLoading ?? false});
     const [sessionEmail] = useOnyx(ONYXKEYS.SESSION, {selector: (session) => session?.email});
+    const [completedHybridAppOnboarding] = useOnyx(ONYXKEYS.NVP_TRYNEWDOT, {selector: onboardingStatusSelector});
+
+    /**
+     * This useEffect tracks changes of `nvp_tryNewDot` value.
+     * We propagate it from OldDot to NewDot with native method due to limitations of old app.
+     */
+    useEffect(() => {
+        if (completedHybridAppOnboarding === undefined) {
+            return;
+        }
+
+        Log.info(`[HybridApp] Onboarding status has changed. Propagating new value to OldDot`, true, {completedHybridAppOnboarding});
+        NativeModules.HybridAppModule.completeOnboarding(completedHybridAppOnboarding);
+    }, [completedHybridAppOnboarding]);
 
     // In iOS, the HybridApp defines the `onReturnToOldDot` event.
     // If we frequently transition from OldDot to NewDot during a single app lifecycle,
